requirements.txt
py-midi
# tkinter ?
pillow
python-rtmidi